requirements.txt
requests
pymongo
mongoengine
python-dateutil
python-daemon
minio
pillow
osmium
elasticsearch
geojson
eventlet
setproctitle
python-slugify